refactor: P1-T0 - 收敛文章抓取站点适配目录
