+++ webclipper/src/collectors/web/article-extract/sites/index.ts
@@ -1,0 +1,13 @@
+export { ARTICLE_FETCH_SITE_SPECS } from '@collectors/web/article-extract/sites/specs';
+export {
+  extractDiscourseOpOnly,
+  findDiscourseOpNode,
+  parseDiscourseTopicPathOnPage,
+} from '@collectors/web/article-extract/sites/discourse';
+export {
+  buildWechatShareMediaGalleryHtml,
+  buildWechatShareMediaGalleryMarkdown,
+  extractWechatShareMediaImageUrls,
+  isWechatShareMediaPage,
+} from '@collectors/web/article-extract/sites/wechat';
+export { isXiaohongshuNotePage, waitForXiaohongshuNoteHydrated } from '@collectors/web/article-extract/sites/xiaohongshu';
