+++ webclipper/src/collectors/web/article-extract/sites/specs.ts
@@ -1,0 +1,1 @@
+export { ARTICLE_FETCH_SITE_SPECS } from '@collectors/web/article-fetch-sites';
